backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.130.0
python-dotenv==1.0.0
